Refactor proof construction code
DRYified the code; by moving repeated methods out into seperate file proofUtils.js

Specifically, moved .computeChallenge() and .parseInputs(). The challenge is computed in exactly the same way for all proofs and I plan to call .parseInputs() for all proofs
