+++ packages/aztec.js/src/proof/joinSplit/index.js
@@ -10,12 +10,12 @@ const { padLeft } = require('web3-utils');
 const extractor = require('./extractor');
 const helpers = require('./helpers');
 const verifier = require('./verifier');
+const proofUtils = require('../proofUtils');
 
 const abiEncoder = require('../../abiEncoder');
 const bn128 = require('../../bn128');
 const Keccak = require('../../keccak');
 const sign = require('../../sign');
-const { K_MAX } = require('../../params');
 
 const { groupReduction } = bn128;
 
@@ -56,85 +56,6 @@ joinSplit.generateBlindingScalars = (n, m) => {
     return scalars;
 };
 
-/**
- * Compute the Fiat-Shamir heuristic-ified challenge variable.
- *   Separated out into a distinct method so that we can stub this for extractor tests
- *
- * @method computeChallenge
- * @memberof proof.joinSplit
- * @param {string} sender Ethereum address of transaction sender
- * @param {string} kPublic public commitment being added to proof
- * @param {number} m number of input notes
- * @param {Object[]} notes array of AZTEC notes
- * @param {Object[]} blindingFactors array of computed blinding factors, one for each note
- */
-joinSplit.computeChallenge = (...challengeVariables) => {
-    const hash = new Keccak();
-
-    const recurse = (inputs) => {
-        inputs.forEach((challengeVar) => {
-            if (typeof (challengeVar) === 'string') {
-                hash.appendBN(new BN(challengeVar.slice(2), 16));
-            } else if (typeof (challengeVar) === 'number') {
-                hash.appendBN(new BN(challengeVar));
-            } else if (BN.isBN(challengeVar)) {
-                hash.appendBN(challengeVar.umod(bn128.curve.n));
-            } else if (Array.isArray(challengeVar)) {
-                recurse(challengeVar);
-            } else if (challengeVar.gamma) {
-                hash.append(challengeVar.gamma);
-                hash.append(challengeVar.sigma);
-            } else if (challengeVar.B) {
-                hash.append(challengeVar.B);
-            } else {
-                throw new Error(`I don't know how to add ${challengeVar} to hash`);
-            }
-        });
-    };
-    recurse(challengeVariables);
-
-    return hash.keccak(groupReduction);
-};
-
-function isOnCurve(point) {
-    const lhs = point.y.redSqr();
-    const rhs = point.x.redSqr().redMul(point.x).redAdd(bn128.curve.b);
-    return (lhs.fromRed().eq(rhs.fromRed()));
-}
-
-/**
- * Validate proof inputs are well formed
- *
- * @method parseInputs
- * @memberof proof.joinSplit
- * @param {Object[]} notes array of AZTEC notes
- * @param {number} m number of input notes
- * @param {string} sender Ethereum address of transaction sender
- * @param {string} kPublic public commitment being added to proof
- */
-joinSplit.parseInputs = (notes, m, sender, kPublic) => {
-    notes.forEach((note) => {
-        if (!note.a.fromRed().lt(bn128.curve.n) || note.a.fromRed().eq(new BN(0))) {
-            throw new Error('viewing key malformed');
-        }
-        if (!note.k.fromRed().lt(new BN(K_MAX))) {
-            throw new Error('note value malformed');
-        }
-        if (note.gamma.isInfinity() || note.sigma.isInfinity()) {
-            throw new Error('point at infinity');
-        }
-        if (!isOnCurve(note.gamma) || !isOnCurve(note.sigma)) {
-            throw new Error('point not on curve');
-        }
-    });
-
-    if (!kPublic.lt(bn128.curve.n)) {
-        throw new Error('kPublic value malformed');
-    }
-    if (m > notes.length) {
-        throw new Error('m is greater than note array length');
-    }
-};
 
 /**
  * Construct AZTEC join-split proof transcript
@@ -195,7 +116,7 @@ joinSplit.constructJoinSplit = (notes, m, sender, kPublic) => {
         };
     });
 
-    const challenge = joinSplit.computeChallenge(sender, kPublicBn, m, notes, blindingFactors);
+    const challenge = proofUtils.computeChallenge(sender, kPublicBn, m, notes, blindingFactors);
 
     const proofData = blindingFactors.map((blindingFactor, i) => {
         let kBar = ((notes[i].k.redMul(challenge)).redAdd(blindingFactor.bk)).fromRed();
